% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-classes.R
\docType{methods}
\name{extract_array}
\alias{extract_array}
\alias{extract_array,GDSArraySeed-method}
\alias{GDSArray-class}
\alias{matrixClass,GDSArray-method}
\alias{GDSMatrix}
\alias{GDSMatrix-class}
\alias{coerce}
\alias{coerce,GDSArray,GDSMatrix-method}
\alias{coerce,GDSMatrix,GDSArray-method}
\alias{coerce,ANY,GDSMatrix-method}
\alias{GDSArray}
\alias{GDSArray-method}
\title{GDSArray constructor and coercion methods.}
\usage{
\S4method{extract_array}{GDSArraySeed}(x, index)

GDSArray(file, name = NA, index = c("list", "IndexList"))
}
\arguments{
\item{x}{the GDSArraySeed object}

\item{index}{the class of the index slot of the
GDSArray. \code{IndexList} is a reference class list defined in
package \code{DelayedArray}, in order to have different
GDSArrays to share the index slot, for a uniform subsetting
performance.}

\item{file}{the gds file name.}

\item{name}{the gds array node to be read into GDSArraySeed / GDSArray. For
\code{GDSArray}, the default value for \code{name} is the
genotype data.}
}
\value{
\code{GDSArray} class object.
}
\description{
GDSArray constructor and coercion methods.

\code{GDSArray}: The function to convert a gds file
    into the GDSArray data structure.
}
\examples{
file <- SNPRelate::snpgdsExampleFileName()
allnodes <- gdsNodes(file)  ## print all available gds nodes in file.
allnodes
GDSArray(file)
GDSArray(file, "sample.annot/pop.group")

file1 <- SeqArray::seqExampleFileName("gds")
allnodes1 <- gdsNodes(file1)  ## print all available gds nodes in file1. 
allnodes1
GDSArray(file1)
GDSArray(file1, "variant.id")
GDSArray(file1, "sample.annotation/family")
GDSArray(file1, "annotation/format/DP/data")
GDSArray(file1, "annotation/info/DP")
}
