% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-methods.R
\docType{methods}
\name{dim,GDSArraySeed-method}
\alias{dim,GDSArraySeed-method}
\alias{dimnames,GDSArraySeed-method}
\alias{gdsfile}
\alias{gdsfile,GDSArraySeed-method}
\alias{gdsfile,GDSArray-method}
\alias{gdsfile,DelayedArray-method}
\alias{gdsfile<-}
\title{GDSArraySeed or GDSArray related methods, slot getters and setters.}
\usage{
\S4method{dim}{GDSArraySeed}(x)

\S4method{dimnames}{GDSArraySeed}(x)

gdsfile(x)

\S4method{gdsfile}{GDSArraySeed}(x)

\S4method{gdsfile}{GDSArray}(x)

\S4method{gdsfile}{DelayedArray}(x)

gdsfile(x) <- value
}
\arguments{
\item{x}{GDSArray, GDSMatrix, GDSArraySeed or SummarizedExperiment
object.}
}
\description{
\code{dim}, \code{dimnames}: dimension and dimnames of
    object contained in the GDS file.

\code{gdsfile}: on-disk location of GDS file
    represented by this object.
}
