% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSlight-class.R
\docType{methods}
\name{GDSlight}
\alias{GDSlight}
\alias{GDSlight-constructor}
\alias{$,GDSlight-method}
\alias{GDSlight-method}
\title{GDSlight constructor and methods.}
\usage{
GDSlight(file, current_path = "")

\S4method{$}{GDSlight}(x, name)
}
\description{
GDSlight is a light-weight class to represent a GDS file. It has the \code{$} method defined as a convenient input for \code{GDSArray} constructor.
}
\details{
\code{$} takes \code{GDSlight} object as input, and output
    by updating \code{@current_path}.
}
