% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdsNode.R
\name{gdsNodes}
\alias{gdsNodes}
\title{gdsNodes}
\usage{
gdsNodes(file)
}
\arguments{
\item{file}{the gds file name.}
}
\value{
a vector of characters.
}
\description{
the function to return all available gds nodes inside
    the input gds file.
}
\examples{
file <- SNPRelate::snpgdsExampleFileName()
gdsNodes(file)
file1 <- SeqArray::seqExampleFileName("gds")
gdsNodes(file1)

}
