% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquireGDS.R
\name{acquireGDS}
\alias{acquireGDS}
\alias{releaseGDS}
\title{Acquire the GDS file connection in R in the \code{gds.class} class.}
\usage{
acquireGDS(path)

releaseGDS(path)
}
\arguments{
\item{path}{String containing a path to a GDS file.}
}
\value{
For \code{acquireGDS}, a \code{gds.class} object, which are
    identical to that returned by \code{gdsfmt::openfn.gds(path)}

For \code{releaseGDS}, any existing \code{gds.class} object for the
\code{path} is disconnected and cleared from cache, and \code{NULL}
is invisibly returned. This is equivalent to that returned by
\code{gdsfmt::closefn.gds()} except it take \code{path} as
input. If \code{path=NULL}, all cached connections are removed.
}
\description{
Acquire a (possibly cached) \code{gds.class} object given it's path.
}
\details{
\code{acquireConn} will cache the \code{gds.class} object
    in the current R session to avoid repeated initialization. This
    improves efficiency for repeated calls. The cached
    \code{gds.class} object for any given \code{path} can be
    deleted by calling \code{releaseGDS} for the same \code{path}.
}
\examples{
fn <- gdsExampleFileName()
gdscon <- acquireGDS(fn)
acquireGDS(fn)  ## just re-uses the cache
releaseGDS(fn)  ## clears the cache

}
\author{
Qian Liu
}
