% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray-classes.R
\docType{class}
\name{GDSArraySeed-class}
\alias{GDSArraySeed-class}
\alias{GDSArraySeed}
\alias{GDSArray}
\title{GDSArraySeed / GDSArray classes and constructors for gds data.}
\usage{
GDSArraySeed(file, name = NA)

GDSArray(file, name = NA, index = c("list", "IndexList"))
}
\arguments{
\item{file}{the gds file name.}

\item{name}{the gds array nodes to be read into GDSArray}

\item{index}{the class of the index slot of the
GDSArray. \code{IndexList} is a reference class list defined in
package \code{DelayedArray}, in order to have different
GDSArrays to share the index slot, for a uniform subsetting
performance.}
}
\description{
GDSArraySeed / GDSArray classes and constructors for gds data.

\code{GDSArraySeed}: The function to generate a GDSArraySeed for the later
    conversion from gds file into GDSArray.

\code{GDSArray}: The function to convert a gds file
    into the GDSArray data structure.
}
