% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFile-class.R
\docType{class}
\name{GDSFile-class}
\alias{GDSFile-class}
\alias{GDSFile}
\alias{GDSFile-constructor}
\alias{gdsfile,GDSFile-method}
\alias{GDSFile-method}
\alias{gdsfile<-,GDSFile-method}
\alias{GDSFile-method}
\alias{GDSFile,gdsfile-method}
\alias{$,GDSFile-method}
\alias{GDSFile-method}
\alias{gdsNodes}
\alias{GDSFile-method}
\alias{gdsNodes,ANY-method}
\alias{gdsNodes,GDSFile-method}
\title{GDSFile constructor and methods.}
\usage{
GDSFile(file, current_path = "")

\S4method{gdsfile}{GDSFile}(x)

\S4method{gdsfile}{GDSFile}(x) <- value

\S4method{$}{GDSFile}(x, name)

\S4method{gdsNodes}{ANY}(x)
}
\arguments{
\item{file}{the GDS file path.}

\item{current_path}{the current path to the closest gds node.}

\item{x}{a \code{GDSFile} object or GDS file path.}

\item{value}{the new gds file path}

\item{name}{the name of gds node}
}
\value{
\code{gdsfile}: the file path of corresponding GDS file.

\code{$}: a \code{GDSFile} with updated \code{@current_path}, or
    \code{GDSArray} object if the \code{current_path} is a valid
    gds node.

\code{gdsNodes}: a character vector for the available gds
    nodes.
}
\description{
\code{GDSFile}: \code{GDSFile} is a light-weight class
    to represent a GDS file. It has the `$` completion method to
    complete any possible gds nodes. If the slot of `current_path`
    in `GDSFile` object represent a valid gds node, it will return
    the `GDSArray` of that node directly. Otherwise, it will return
    the `GDSFile` object with an updated `current_path`.

the \code{GDSFile} constructor.

\code{gdsfile}: \code{file} slot getter and setter for
    \code{GDSFile} object.

\code{gdsNodes}: to get the available gds nodes from
    the \code{GDSFile} object or the file path with extension of
    ".gds".
}
\examples{
file <- SNPRelate::snpgdsExampleFileName()
gdsNodes(file)
file1 <- SeqArray::seqExampleFileName("gds")
gdsNodes(file1)
gf <- GDSFile(file)
gdsNodes(gf)
gdsfile(gf)
}
