% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFile-class.R
\docType{methods}
\name{GDSFile}
\alias{GDSFile}
\alias{GDSFile-constructor}
\alias{$,GDSFile-method}
\alias{GDSFile-method}
\alias{gdsfile,GDSFile-method}
\alias{GDSFile-method}
\alias{gdsNodes}
\alias{GDSFile-method}
\title{GDSFile constructor and methods.}
\format{An object of class \code{NULL} of length 0.}
\usage{
GDSFile(file, current_path = "")

\S4method{$}{GDSFile}(x, name)

\S4method{gdsfile}{GDSFile}(x)
}
\value{
a character vector for the available gds nodes.
}
\description{
GDSFile is a light-weight class to represent a GDS
    file. It has the \code{$} method defined as a convenient input
    for \code{GDSArray} constructor.

\code{file} slot getter for \code{GDSFile} object.

to get the available gds nodes for the \code{GDSFile} object.
}
\details{
\code{$} takes \code{GDSFile} object as input, and output
    by updating \code{@current_path}.
}
\keyword{datasets}
