% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSFile-class.R
\docType{class}
\name{GDSFile-class}
\alias{GDSFile-class}
\alias{GDSFile}
\alias{GDSFile-constructor}
\alias{gdsfile,GDSFile-method}
\alias{GDSFile-method}
\alias{gdsfile<-,GDSFile-method}
\alias{GDSFile-method}
\alias{GDSFile,gdsfile-method}
\alias{$,GDSFile-method}
\alias{GDSFile-method}
\alias{gdsnodes}
\alias{GDSFile-method}
\alias{gdsnodes,ANY-method}
\alias{gdsnodes,GDSFile-method}
\title{GDSFile constructor and methods.}
\usage{
GDSFile(file, current_path = "")

\S4method{gdsfile}{GDSFile}(x)

\S4method{gdsfile}{GDSFile}(x) <- value

\S4method{$}{GDSFile}(x, name)

\S4method{gdsnodes}{ANY}(x)
}
\arguments{
\item{file}{the GDS file path.}

\item{current_path}{the current path to the closest gds node.}

\item{x}{a \code{GDSFile} object. or GDS file path (for
\code{gdsnodes()}).}

\item{value}{the new gds file path}

\item{name}{the name of gds node}
}
\value{
\code{gdsfile}: the file path of corresponding GDS file.

\code{$}: a \code{GDSFile} with updated \code{@current_path}, or
    \code{GDSArray} object if the \code{current_path} is a valid
    gds node.

\code{gdsnodes}: a character vector for the available gds
    nodes. When input is GDS file path, it returns all available
    gds nodes within the GDS file, no matter there is value or
    not. When input is \code{GDSFile} object, it returns only the
    gds nodes that could construct unique \code{GDSArray} objects,
    which means that the gds node has non-zero-dimensions, and is
    actually array, and all \code{GDSArray}s returned from these
    nodes are unique (by excluding the gds nodes that has `code{~}
    prefix).
}
\description{
\code{GDSFile}: \code{GDSFile} is a light-weight class
    to represent a GDS file. It has the `$` completion method to
    complete any possible gds nodes. If the slot of `current_path`
    in `GDSFile` object represent a valid gds node, it will return
    the `GDSArray` of that node directly. Otherwise, it will return
    the `GDSFile` object with an updated `current_path`.

\code{GDSFile}: the \code{GDSFile} class constructor.

\code{gdsfile}: \code{file} slot getter for
    \code{GDSFile} object.

\code{gdsfile<-}: \code{file} slot setter for
    \code{GDSFile} object.

\code{gdsnodes}: to get the available gds nodes from
    the \code{GDSFile} object or the file path with extension of
    ".gds".
}
\examples{
file <- SNPRelate::snpgdsExampleFileName()
gdsnodes(file)
file1 <- SeqArray::seqExampleFileName("gds")
gdsnodes(file1)
gf <- GDSFile(file)
gdsnodes(gf)
gdsfile(gf)
}
